Introduced Editor Mode
Prompt: I would like to introduce an editor mode vs game mode. The game should load the data as normal, but the objects should not be interactable by default. Place a checkbox in the top left corner which should be used to enable editor mode. This should set a flag which enables editing of the object position and sizes.

Then:
There is a bug in the code: MOUSEBUTTONDOWN events are not passed down into the editor when editor mode is enabled.