Added point and click to find objects mechanic
Prompt: When not in editor mode (i.e., game mode), the player is asked to find all of the items hidden in the picture. When they click on an item, it should animate and disappear, and a counter of how many items remain should be deremented. When the player has found them all, it should display "Well done!" and a Restart button should be shown, which resets the game back to its initial state.