Added wiggle hint if player hasn't found an item for a while
Prompts:
Add a timer to the UI. If the player hasn't found an object for around 20-30 seconds, select one unfound item and give it a wiggle animation

Move the wiggle animation logic inside the scene object class

Also reset the wiggle interval after triggering a wiggle

Reset the wiggle item key too. Move this reset logic into a helper function.

Actually we don't need to remember the wiggle item key - it can be a fire-and-forget mechanism. Just remember the wiggle interval.