Added inventory section
Prompts:
Let's add an inventory section. Extend the window by 100 pixels underneath the background. Leave it with a black background. Whenever the player finds an item, add an image of the same object in the inventory.

There isn't enough space to render all of the items in the inventory. Divide the inventory area into a grid of two rows and however many columns can fit. Render each found item in the next available cell.

Now move the hard coded constants into our CONSTANTS file. Also move the wiggle timer range, the size and position of the restart button, and any other constants that are useful to configure.